initial working derivation
