+++ users/dots/nvim/lsp/nix.lua
@@ -1,0 +1,23 @@
+---@type vim.lsp.Config
+return {
+  name = "nil",
+  cmd = { "nil" },
+  filetypes = { "nix" },
+  root_markers = { "flake.nix", ".git" },
+  settings = {
+    ["nil"] = {
+      nix = {
+        binary = "nix",
+        maxMemoryMB = nil,
+        flake = {
+          autoEvalInputs = false,
+          autoArchive = false,
+          nixpkgsInputName = "nixpkgs",
+        },
+      },
+      formatting = {
+        command = { "alejandra", "--quiet" },
+      },
+    },
+  },
+}
