+++ users/dots/nvim/lua/plugins/extraft.lua
@@ -1,0 +1,5 @@
+vim.filetype.add {
+  pattern = {
+    ['.*/hyprland/.*%.conf'] = 'hyprlang',
+  },
+}
