+++ users/dots/nvim/lua/plugins/init.lua
@@ -1,0 +1,11 @@
+require('plugins.options')
+require('plugins.keybinds')
+require('plugins.extraft')
+require('plugins.themes')
+require('plugins.generic')
+require('plugins.treesitter')
+require('plugins.telescope')
+require('plugins.cmp')
+require('plugins.gitsigns')
+require('plugins.dbee')
+require('plugins.lsp')
